% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignerTrimmer.R
\name{alignerTrimmer}
\alias{alignerTrimmer}
\title{alignerTrimmer}
\usage{
alignerTrimmer(path.to.fasta = getwd(), samples.cutoff = 10,
  quantile.cutoff = 0.25, segsites.cutoff = 0.5, remove.gaps = T,
  input.id.name = "uce-", output.prefix = "ALTRIM_",
  output.dir = "OUT")
}
\arguments{
\item{path.to.fasta}{character string. path to the directory of fasta files.}

\item{samples.cutoff}{integer. Exclude loci that have less samples than the sample.cutoff.}

\item{quantile.cutoff}{float. Exclude all sequences with length bellow the quantile 
of the distribution of lengths across all samples. This is to exclude samples with short sequences before trimming.}

\item{segsites.cutoff}{float. Exclude locus if the number segsites is larger than the segsites.cutoff of total length of the alignment.}

\item{remove.gaps}{logical. If TRUE remove gaps and trim alignment.}

\item{input.id.name}{character string. Text indicating a wildcard to find the loci to process.}

\item{output.prefix}{character string. Text prefix to identify the output.}

\item{output.dir}{character string. Path to the output directory.}
}
\value{
processed sequence alignmnents to the to the output directory.
}
\description{
This function aligns and trims all fasta sequences present in a directory
}
\author{
Marcelo Gehara and Edward Myers
}
