% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excludeFromFasta.R
\name{excludeFromFasta}
\alias{excludeFromFasta}
\title{excludeFromFasta}
\usage{
excludeFromFasta(path, samples2exclude, out.prefix.name = "out")
}
\arguments{
\item{path}{character string. Path to the fasta alignments}

\item{samples2exclude}{one column data frame with the name of the samples to exclude from fasta alignments.}

\item{out.prefix.name}{character string. A character string to use as prefix of new fasta alignments. default is "out"}
}
\description{
this is to remove samples from final fasta alignments
}
\author{
Marcelo Gehara
}
