% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta2strc.R
\name{fasta2struc}
\alias{fasta2struc}
\title{fasta2struc}
\usage{
fasta2struc(path, samples2include, write.output = TRUE)
}
\arguments{
\item{path}{path to the directory where all fasta alignments are.}

\item{samples2include}{one column data frame with the name of the samples to include in the first column.}

\item{write.output}{logical. If TRUE writes the structure file to working directory.}
}
\value{
Structure-like genotype matrix.
}
\description{
THIS SCRIPT PULLS OUT A RANDOM SNP FROM THE ALIGNMENTS TO MAKE A STRUCTURE FILE.
}
\author{
Gehara, M.
}
