% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta2vcf.R
\name{fasta2VCF}
\alias{fasta2VCF}
\title{fasta2VCF}
\usage{
fasta2VCF(path, samples2include)
}
\arguments{
\item{path}{path to the directory where all fasta alignments are.}

\item{samples2include}{one column data frame with the name of the samples to include in the first column}
}
\value{
out.vcf file to the path directory
}
\description{
Converts a colection of phased fasta alignments into a VCF file
}
\author{
Gehara, M
}
