% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findReplaceNames.R
\name{findReplaceNames}
\alias{findReplaceNames}
\title{findReplaceNames}
\usage{
findReplaceNames(path.to.fasta, input.id.name,
  expr.to.replace = "(_uce-\\\\d+)", replacement = "")
}
\arguments{
\item{path.to.fasta}{character string. path to the directory of fasta files.}

\item{input.id.name}{character string. Text indicating a wildcard to find the loci to process.}

\item{expr.to.replace}{charactees to exclude from samples names}

\item{replacement}{}
}
\value{
processed sequence alignmnents to the to the output directory.
}
\description{
This script will find and replace expressions in sequence names of fasta files.
}
\details{
Ex. if the names starts as DBS_355_uce-8046a, it will end up as DBS_355a (note the regex in 'gsub')
}
\author{
Edward Myers and Marcelo Gehara
}
