% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.filter.R
\name{str.filter}
\alias{str.filter}
\title{str.filter}
\usage{
\method{str}{filter}(structure.file, cutoff = NULL, filter.loci = T,
  filter.samp = F)
}
\arguments{
\item{structure.file}{a matrix or data frame object representing the structure-like genotype file}

\item{cutoff}{float. A percentage cutoff between 0 and 1 to exclude loci or individuals with missing data. Defaut NULL.}

\item{filter.loci}{logical. If TRUE loci with missing samples will be filtered out.}

\item{filter.samp}{logical. If TRUE samples with missing loci will be filtered out.}
}
\description{
this function filters out individuals or loci with missing data according to a cut off percentage
}
\author{
Marcelo Gehara
}
